feat(example): rename project to be unique
