+++ example/.gitignore
@@ -58,7 +58,3 @@ buck-out/
 
 # CocoaPods
 /ios/Pods/
-
-# yalc
-.yalc/
-yalc.lock
