+++ example/__tests__/App-test.tsx
@@ -9,7 +9,6 @@ import App from '../App';
 // Note: test renderer must be required after react-native.
 import renderer from 'react-test-renderer';
 
-// eslint-disable-next-line jest/no-disabled-tests
-it.skip('renders correctly', () => {
+it('renders correctly', () => {
   renderer.create(<App />);
 });
