+++ example/android/app/_BUCK
@@ -35,12 +35,12 @@ android_library(
 
 android_build_config(
     name = "build_config",
-    package = "com.example",
+    package = "com.jiterarnuilibraryexample",
 )
 
 android_resource(
     name = "res",
-    package = "com.example",
+    package = "com.jiterarnuilibraryexample",
     res = "src/main/res",
 )
