+++ example/android/app/src/main/AndroidManifest.xml
@@ -1,11 +1,7 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-  package="com.example">
+  package="com.jiterarnuilibraryexample">
 
     <uses-permission android:name="android.permission.INTERNET" />
-    <!-- Added permissions -->
-    <uses-permission android:name="android.permission.CAMERA" />
-    <uses-permission android:name="android.permission.READ_EXTERNAL_STORAGE" />
-    <uses-permission android:name="android.permission.WRITE_EXTERNAL_STORAGE" />
 
     <application
       android:name=".MainApplication"
@@ -24,10 +20,6 @@
             <action android:name="android.intent.action.MAIN" />
             <category android:name="android.intent.category.LAUNCHER" />
         </intent-filter>
-      </activity>
-      <activity
-        android:name="com.theartofdev.edmodo.cropper.CropImageActivity"
-        android:theme="@style/Base.Theme.AppCompat">
       </activity>
     </application>
 </manifest>
