+++ example/android/app/src/main/java/com/jiterarnuilibraryexample/MainActivity.java
@@ -1,7 +1,6 @@
-package com.example;
+package com.jiterarnuilibraryexample;
 import expo.modules.ReactActivityDelegateWrapper;
 
-import android.os.Bundle;
 import com.facebook.react.ReactActivity;
 
 import com.facebook.react.ReactActivityDelegate;
@@ -10,27 +9,22 @@ import com.swmansion.gesturehandler.react.RNGestureHandlerEnabledRootView;
 
 public class MainActivity extends ReactActivity {
 
-  @Override
-  protected void onCreate(Bundle savedInstanceState) {
-    super.onCreate(null);
-  }
-
   /**
    * Returns the name of the main component registered from JavaScript. This is used to schedule
    * rendering of the component.
    */
   @Override
   protected String getMainComponentName() {
-    return "example";
+    return "JiteraRNUILibraryExample";
   }
 
   @Override
-    protected ReactActivityDelegate createReactActivityDelegate() {
-        return new ReactActivityDelegate(this, getMainComponentName()) {
-            @Override
-            protected ReactRootView createRootView() {
-                return new RNGestureHandlerEnabledRootView(MainActivity.this);
-            }
-        };
-    }
+  protected ReactActivityDelegate createReactActivityDelegate() {
+      return new ReactActivityDelegateWrapper(this, new ReactActivityDelegate(this, getMainComponentName()) {
+          @Override
+          protected ReactRootView createRootView() {
+              return new RNGestureHandlerEnabledRootView(MainActivity.this);
+          }
+      });
+  }
 }
