+++ example/app.json
@@ -1,4 +1,4 @@
 {
-  "name": "example",
-  "displayName": "example"
+  "name": "JiteraRNUILibraryExample",
+  "displayName": "JiteraRNUILibraryExample"
 }
