+++ example/ios/JiteraRNUILibraryExample/AppDelegate.m
@@ -33,7 +33,7 @@ static void InitializeFlipper(UIApplication *application) {
 
   RCTBridge *bridge = [self.reactDelegate createBridgeWithDelegate:self launchOptions:launchOptions];
   RCTRootView *rootView = [self.reactDelegate createRootViewWithBridge:bridge
-                                                   moduleName:@"example"
+                                                   moduleName:@"JiteraRNUILibraryExample"
                                             initialProperties:nil];
 
   if (@available(iOS 13.0, *)) {
