+++ example/ios/JiteraRNUILibraryExample/Images.xcassets/Contents.json
