+++ example/ios/JiteraRNUILibraryExampleTests/Info.plist
