+++ example/ios/JiteraRNUILibraryExampleTests/JiteraRNUILibraryExampleTests.m
@@ -7,11 +7,11 @@
 #define TIMEOUT_SECONDS 600
 #define TEXT_TO_LOOK_FOR @"Welcome to React"
 
-@interface exampleTests : XCTestCase
+@interface JiteraRNUILibraryExampleTests : XCTestCase
 
 @end
 
-@implementation exampleTests
+@implementation JiteraRNUILibraryExampleTests
 
 - (BOOL)findSubviewInView:(UIView *)view matching:(BOOL(^)(UIView *view))test
 {
