+++ example/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "example",
+  "name": "jiterarnuilibraryexample",
   "version": "0.0.1",
   "private": true,
   "scripts": {
@@ -7,41 +7,40 @@
     "ios": "react-native run-ios",
     "start": "react-native start",
     "test": "jest",
-    "lint": "eslint . --ext .js,.jsx,.ts,.tsx",
-    "add-library": "yarn cache clean @jitera/jitera-rn-ui-library && yarn add ../",
-    "reset": "watchman watch-del-all"
+    "lint": "eslint . --ext .js,.jsx,.ts,.tsx"
   },
   "dependencies": {
-    "@jitera/jitera-rn-ui-library": "../",
+    "@jitera/jitera-rn-ui-library": "../jitera-rn-ui-library/",
     "@react-native-community/datetimepicker": "^5.1.0",
-    "@react-navigation/native": "^6.0.6",
-    "@react-navigation/native-stack": "^6.2.5",
-    "deepmerge": "^4.2.2",
+    "@react-navigation/native": "^6.0.8",
+    "@react-navigation/native-stack": "^6.5.0",
     "expo": ">=44.0.0-0 <45.0.0",
-    "expo-image-picker": "~12.0.1",
+    "expo-auth-session": "3.5.0",
+    "expo-image-picker": "^12.0.2",
+    "expo-random": "^12.1.2",
     "react": "17.0.2",
-    "react-native": "0.67.1",
+    "react-native": "0.67.3",
     "react-native-gesture-handler": "^1.10.3",
-    "react-native-reanimated": "~2.2.4",
-    "react-native-safe-area-context": "^3.3.2",
-    "react-native-screens": "^3.10.2",
+    "react-native-reanimated": "2.2.4",
+    "react-native-safe-area-context": "^4.1.2",
+    "react-native-screens": "^3.13.1",
     "react-native-scroll-bottom-sheet": "^0.7.0",
-    "react-native-webview": "^11.17.0"
+    "react-native-webview": "^11.17.2"
   },
   "devDependencies": {
     "@babel/core": "^7.12.9",
     "@babel/runtime": "^7.12.5",
     "@react-native-community/eslint-config": "^2.0.0",
     "@types/jest": "^26.0.23",
-    "@types/react-native": "^0.66.4",
+    "@types/react-native": "^0.66.15",
     "@types/react-test-renderer": "^17.0.1",
     "@typescript-eslint/eslint-plugin": "^5.7.0",
     "@typescript-eslint/parser": "^5.7.0",
     "babel-jest": "^26.6.3",
-    "babel-plugin-module-resolver": "^4.1.0",
     "eslint": "^7.14.0",
     "jest": "^26.6.3",
     "metro-react-native-babel-preset": "^0.66.2",
+    "pod-install": "^0.1.32",
     "react-test-renderer": "17.0.2",
     "typescript": "^4.4.4"
   },
